+++ src/utils/debouncer.js
@@ -4,7 +4,7 @@
  * @param {Function} callback - The callback function
  * @param {Number} delay - The delay number value
  *
- * @returns {Function} - The callback function wrapped in `setTimeout` function
+ * @return {Function} - The callback function wrapped in `setTimeout` function
  */
 export default (callback, delay) => {
   let inDebounce;