+++ src/utils/eventEmitter.js
@@ -5,6 +5,7 @@
  * @param {Object} detail - The detail object containing relevant information or null
  * @param {String} name - The name of the event fired
  * 
+ * @return {void}
  */
 export default (target, detail, name) => {
   // Dispatch event on input